app/requirements.txt
# Core dependencies
python-telegram-bot==13.15
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Optimization and monitoring
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
tabulate==0.9.0
tqdm==4.66.1

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

aiogram>=2.20.0
aiohttp>=3.8.1
aioschedule>=0.5.2
aiopayok>=0.2.1
translatepy>=2.3
pillow>=9.0.0
gtts>=2.2.3
requests>=2.27.1
websocket-client>=1.6.4
python-socketio>=5.11.1 